contend with rate exceeded errors
the api gives us information about how many calls we’ve made and how close we are to the limit, but it doesn’t give us a way to know when the time was started.  The will rescue from a RATE_EXCEEDED error and wait for a set period and retry the operation until it doesn’t encounter the RATE_EXCEEDED error.